Update area and circumference of circle.py

+++ basics/day2-practice/area and circumference of circle.py
@@ -1,4 +1,4 @@
-r=float(input("Radius of circle:"));
+r=int(input("Radius of circle:"));
 A=r*r*3.1416;
 C=2*3.1416*r;
 print("Area of circle=",A);
